+++ database/migrations/2019_04_01_195330_bigcommerce_create_product_reviews_table.php
@@ -14,8 +14,8 @@ class BigCommerceCreateProductReviewsTable extends Migration
     public function up()
     {
         Schema::create('bigcommerce_product_reviews', function (Blueprint $table) {
-            $table->unsignedInteger('id');
-            $table->unsignedInteger('product_id');
+            $table->unsignedBigInteger('id');
+            $table->unsignedBigInteger('product_id');
             $table->string('name');
             $table->string('title');
             $table->string('text');
