+++ database/migrations/2019_04_03_195360_bigcommerce_create_product_options_table.php
@@ -14,8 +14,8 @@ class BigCommerceCreateProductOptionsTable extends Migration
     public function up()
     {
         Schema::create('bigcommerce_product_options', function (Blueprint $table) {
-            $table->unsignedInteger('id');
-            $table->unsignedInteger('product_id');
+            $table->unsignedBigInteger('id');
+            $table->unsignedBigInteger('product_id');
             $table->string('name');
             $table->string('display_name');
             $table->string('type');
